requirements.txt
Flask==3.0.0
PySastrFlask==2.3.3
python-docx==0.8.11
pypdf==3.15.0
pytesseract==0.3.10
Pillow==10.0.0
pdf2image==1.16.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
python-dotenv==1.0.0
Werkzeug>=2.0.0
